requirements.txt
# Runtime requirements
boto3==1.14.12
sagemaker==2.1.0
pathlib2==2.3.5
pyyaml==5.4.1
mypy-extensions==0.4.3
kubernetes==24.2.0
awscli==1.25.22